usv_system/visualization/requirements.txt
Flask==2.0.2
Flask-SocketIO==5.1.1
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
PyYAML>=6.0
eventlet==0.33.0
python-engineio==4.3.1
python-socketio[client]>=5.5.0
requests>=2.25.0 